+++ include/heembp_param.hpp
@@ -2,6 +2,8 @@
 #ifndef HEEMBP_INCLUDE_HEEMBP_PARAM
 #define HEEMBP_INCLUCE_HEEMBP_PARAM
 
+#include <cstdint>
+
 namespace heembp_param {
     //This project(HEEModByPuls)'s Parameter without COORDINATE PARAMETER
     // if you want to customaize Coordinate parameter you should look at "/include/coordinatespec.hpp"
@@ -10,6 +12,10 @@ namespace heembp_param {
     // Don't use Pointer type constant
 
     constexpr double PI = 3.1415926535;
+    constexpr std::int_fast8_t kRealDimensionNum = 1;
+    constexpr std::int_fast8_t kVelocityDimensionNum = 2;
+    constexpr std::int_fast16_t kRealGridNum[kRealDimensionNum] ={128}; 
+    constexpr std::int_fast16_t kVelocityGridNum[kVelocityDimensionNum]={32,32};
 
 
 }
