+++ src/coordinate_spec.cpp
@@ -1,0 +1,15 @@
+#include <cstdint>
+#include "../include/coordinate_spec.hpp"
+
+
+namespace coordinate_spec{ 
+    CoordinateSpec::CoordinateSpec(){
+
+    }
+
+    std::int_fast32_t CoordinateSpec::calcTotalGridNum(){
+
+    }
+
+
+}
